+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Changed "Golden Slime Rain" to despawn healthy slimes when the effect is over.
+- Changed "Give Item" effects to prefix the item name with the viewer's name.
 
 ## [1.0.8] - 2022-10-25